+++ common/data/src/main/java/org/thingsboard/server/common/data/cf/configuration/aggregation/AggMetric.java
@@ -1,5 +1,5 @@
 /**
- * Copyright © 2016-2025 The Thingsboard Authors
+ * Copyright © 2016-2026 The Thingsboard Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -27,6 +27,6 @@ public class AggMetric {
     private AggFunction function;
     private String filter;
     private AggInput input;
-    private Long defaultValue;
+    private Double defaultValue;
 
 }
